+++ CHANGELOG.md
@@ -8,6 +8,8 @@
 - Added "Current Elapsed Time" to document execution time prior to dialog creation
 - Improved `quitScript` function to immediately exit the script when the user clicks "Close"
 - Added "set -x" when `operationMode` is set to "test" (to better identify variable initialization issues; I'm looking at you, SSID!)
+- (Hopefully) improved regex for "Palo Alto Networks GlobalProtect VPN IP address" to avoid "JSON import failed" error
+- Corrected Slack Webhook (thanks, @drtaru!)
 
 ### 2.0.0 (17-Jul-2025)
 - Renamed to "Mac Health Check" (thanks, @uurazzle and @scriptingosx!)